vassar-feetech-servo-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vassar-feetech-servo-sdk"
version = "0.1.0"
description = "Vassar College's Python SDK for Feetech servo control and position reading"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Vassar Robotics", email = "hello@vassarrobotics.com"},
]
maintainers = [
    {name = "Vassar Robotics", email = "hello@vassarrobotics.com"},
]
keywords = ["robotics", "servo", "feetech", "serial", "hardware"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=0.990",
    "build>=0.9",
    "twine>=4.0",
]

[project.urls]
Homepage = "https://github.com/vassar-robotics/feetech-servo-sdk"
Repository = "https://github.com/vassar-robotics/feetech-servo-sdk"
Issues = "https://github.com/vassar-robotics/feetech-servo-sdk/issues"

[project.scripts]
vassar-servo = "vassar_feetech_servo_sdk.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["vassar_feetech_servo_sdk*", "scservo_sdk*"]

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true